refactor(anti-fraud service): Set right ubication of components
